stenosis-detection/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
torch==2.4.0
torchvision==0.19.0
ultralytics==8.2.50
opencv-python==4.10.0.84
scikit-image==0.24.0
aiofiles==24.1.0
pydantic==2.8.2
